// src/types/interface/movie/interface.series.ts
export interface ISeriesPayload {
  title: string;
  description: string;
  director?: string;
  categoryId: string;
  cast: string;
  isPremium: string;
  price: string;
  userId?: string;
  releaseDate?: string;
  genre?: string;
  poster?: string;
  posterUrlPublicId?: string;
  trailer?: string;
  trailerUrlPublicId?: string;
}